+++ ICD.Connect.Telemetry.CrestronPro/Devices/FusionRoomAdapter.cs
@@ -3,15 +3,16 @@ using System.Collections.Generic;
 using ICD.Common.Logging.LoggingContexts;
 using ICD.Common.Utils;
 using ICD.Common.Utils.EventArguments;
+using ICD.Common.Utils.Timers;
 using ICD.Connect.Misc.CrestronPro.Utils;
 using ICD.Connect.Settings;
 using ICD.Connect.Telemetry.Nodes;
 #if SIMPLSHARP
 using Crestron.SimplSharpPro;
 using Crestron.SimplSharpPro.Fusion;
+using Crestron.SimplSharpPro.CrestronThread;
 using ICD.Connect.Misc.CrestronPro;
 using System.Linq;
-using ICD.Connect.Misc.CrestronPro.Extensions;
 using ICD.Connect.Telemetry.CrestronPro.Assets;
 using ICD.Common.Properties;
 using ICD.Common.Utils.Extensions;
@@ -50,6 +51,7 @@ namespace ICD.Connect.Telemetry.CrestronPro.Devices
 
 #if SIMPLSHARP
 		private FusionRoom m_FusionRoom;
+		private Thread m_RviGenerationThread;
 #endif
 
 		#region Properties
@@ -194,6 +196,10 @@ namespace ICD.Connect.Telemetry.CrestronPro.Devices
 
 #if SIMPLSHARP
 			SetFusionRoom(null);
+
+			Thread thread = m_RviGenerationThread;
+			if (thread != null)
+				thread.Abort();
 #endif
 		}
 
@@ -493,12 +499,32 @@ namespace ICD.Connect.Telemetry.CrestronPro.Devices
 		public void RebuildRvi()
 		{
 #if SIMPLSHARP
-			m_FusionRoom.ReRegister();
-			FusionRVI.GenerateFileForAllFusionDevices();
+			m_RviGenerationThread = new Thread(RviGenerationThreadCallback, null)
+			{
+				Name = "RviGenerationThread",
+				Priority = Thread.eThreadPriority.LowestPriority
+			};		
 #else
 			throw new NotSupportedException();
 #endif
 		}
+
+#if SIMPLSHARP
+		private object RviGenerationThreadCallback(object userspecific)
+		{
+			Logger.Log(eSeverity.Debug, "Starting RVI Generation");
+			IcdStopwatch stopwatch = new IcdStopwatch();
+			
+			stopwatch.Start();          
+			FusionRVI.GenerateFileForAllFusionDevices();
+			stopwatch.Stop();
+			Logger.Log(eSeverity.Debug, "RVI Generation took {0}ms", stopwatch.ElapsedMilliseconds);
+			
+			m_RviGenerationThread = null;
+			
+			return null;
+		}
+#endif
 
 		#endregion